refactor: organize SiteConfig mappers
