+++ src/lib/checks/index.ts
@@ -1,0 +1,1 @@
+export * from "./is-record";
