+++ src/lib/checks/is-record.ts
@@ -1,0 +1,3 @@
+export function isRecord(v: unknown): v is Record<string, unknown> {
+	return typeof v === "object" && v !== null && !Array.isArray(v);
+}
