+++ src/lib/content/models/index.ts
@@ -1,0 +1,2 @@
+export * from "./link";
+export * from "./site-config";
