+++ src/lib/content/models/site-config.ts
@@ -1,0 +1,7 @@
+import type { Link } from "./link";
+
+export interface SiteConfig {
+	siteTitle: string;
+	footerText: string;
+	links: Link[];
+}
